+++ programs/config/Cargo.toml
@@ -17,6 +17,10 @@ serde_derive = "1.0.103"
 solana-logger = { path = "../../logger", version = "0.22.0" }
 solana-sdk = { path = "../../sdk", version = "0.22.0" }
 
+[dev-dependencies]
+solana-runtime = { path = "../../runtime", version = "0.22.0" }
+assert_matches = "1.3.0"
+
 [lib]
 crate-type = ["lib", "cdylib"]
 name = "solana_config_program"
