--- programs/config_tests/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "solana-config-tests"
-version = "0.22.0"
-description = "Solana config api tests"
-authors = ["Solana Maintainers <maintainers@solana.com>"]
-repository = "https://github.com/solana-labs/solana"
-license = "Apache-2.0"
-homepage = "https://solana.com/"
-edition = "2018"
-
-[dependencies]
-bincode = "1.2.0"
-log = "0.4.8"
-serde = "1.0.103"
-serde_derive = "1.0.103"
-solana-logger = { path = "../../logger", version = "0.22.0" }
-solana-sdk = { path = "../../sdk", version = "0.22.0" }
-solana-config-program = { path = "../config", version = "0.22.0" }
-
-
-[dev-dependencies]
-solana-runtime = { path = "../../runtime", version = "0.22.0" }
-assert_matches = "1.3.0"
