+++ programs/storage/Cargo.toml
@@ -19,6 +19,9 @@ serde_derive = "1.0.103"
 solana-logger = { path = "../../logger", version = "0.22.0" }
 solana-sdk = { path = "../../sdk", version = "0.22.0" }
 
+[dev-dependencies]
+assert_matches = "1.3.0"
+
 [lib]
 crate-type = ["lib", "cdylib"]
 name = "solana_storage_program"
